backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Cors
Flask-Bcrypt
Flask-JWT-Extended
requests
psycopg2-binary
python-dotenv
google-genai
PyPDF2
